Remove launcher and simple shell
Maintaining a simple desktop shell cost time that is better spent
on the actual Liri OS shell.

Also, the launcher now is not needed since we want to run compositors
under our QPA plugin and provide support for X11 and Wayland
as an EGL device integration.

Remove CLI screencaster since we already have one on liri-workspace.