Fix build with libdrm 2.4.78
libdrm 2.4.78 bumped the ABI from 2 to 3 and added a new function
pointer page_flip_handler2 to the end of the structure.

Initialize the struct and fill only the fields we are interested in,
set the version explicitly to 2 instead of DRM_EVENT_CONTEXT_VERSION
to make sure the library knows we are using an older version and don't
call the new function pointer.

See also: https://blogs.s-osg.org/libdrm-event-handling-youre-probably-wrong/