xwayland: Let Xwayland allocate a display for us
Rewrite the Xwayland server startup code to make it allocate a display
for us so we don't need to create the lock file and figure out what
display it should run.

Automatic startup and socket activation code were also removed,
now the compositor needs to explicitely start the server when
it's ready so we are sure that screens were created.

Issue: #10