Rebase QPA plugin on new eglfs
Upstream eglfs has advanced so much that we now need to rebase our
code on top of it.

The new code is based on the future Qt 5.10 (current dev branch) that
has already a good portion of our changes ported to it, especially the
EDID parser.

The new eglfs supports NVIDIA proprietary drivers which is very
important to us, but it still doesn't support logind nor setting
modes and a few improvements for mouse cursor position and enter/leave
events that we submitted for review.