Screen management overhaul
Screen management is complex but it doesn't need to be.

Replace ScreenManager and its backends with ScreenModel, a model that
is fed by either a JSON configuration file or real screen information
provided by QGuiApplication.

Upon component completion the model is fed with data, each item is
a ScreenItem, that is an object that expose screen properties.

Since Qt 5.9 we can access manufacturer and model from QScreen and
modes from QPlatformScreen because we upstreamed that code.
This means we no longer need to use our EglFSScreen thus making the
code work with other QPA plugins.

Starting with Qt 5.9 we can assign a screen to a window from QML,
this means that most of QuickOutput is now useless.

All integration between the output window and Screencaster has
been moved to it.  QuickOutput now just offers a modes list
property so that QML compositors can add modes from ScreenModel.

Remove dependency between LiriWaylandServer and LiriEglFS since
we no longer need to access our custom EglFSScreen, instead we
can now access QPlatformScreen directly.

Closes: #30